- Integrated PaymentService for processing payments based on user and…

+++ src/main/java/com/example/Rental/models/Entity/Payment.java
@@ -8,10 +8,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 import java.time.LocalDateTime;
 
 @Data
 @Builder
+
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
@@ -30,7 +32,7 @@ public class Payment {
     private PaymentMethod paymentMethod;
 
     @Column(name = "amount", nullable = false)
-    private Double amount;
+    private BigDecimal amount;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
